Make randpkt generate the correct IP version for IPv6
IP version is always 6 and as a consequence of setting the
first byte to 0x60 the 4 high-order bits for the
traffic class field are set to zero.

Otherwise the IPv6 dissector does not look past the first TVB
byte, making randpkt useless to exercise the dissector code.

Change-Id: I372ab7f71e6c972106f9dd46edec642ca53b9557
Reviewed-on: https://code.wireshark.org/review/16708
Petri-Dish: João Valverde <j@v6e.pt>
Tested-by: João Valverde <j@v6e.pt>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: João Valverde <j@v6e.pt>