checklicenses: Fix a FIXME
We do not whitelist GPLv2 only code because of the
incompatibility with GPLv3.

Move the path exception to its proper place. These files are
dual-licensed in the sense of "choose one or the other", not
"both licenses are in use".

If the exception becomes unnecessary later it can be removed,
but let's get rid of the FIXME until then.

Change-Id: Ia040284b72c2ff6588b0544896cc37ae19a61613
Reviewed-on: https://code.wireshark.org/review/16957
Reviewed-by: João Valverde <j@v6e.pt>
(cherry picked from commit e3009b0b1c78b25337c915fb66232314f769688c)
Reviewed-on: https://code.wireshark.org/review/17264
Petri-Dish: Pascal Quantin <pascal.quantin@gmail.com>
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>