WiX: Disable desktop icon and shortcut checkboxes.
Enable and disable the Wireshark and Wireshark-gtk desktop icon and
shortcut checkboxes depending on our feature selections to match the
NSIS installer behavior (and make sense).

Set the level of "Fe.WiresharkGTK" to 2 so that it's disabled by default.
This roughly matches the behavior of the NSIS installer.

Note that we can't enable or disable radio buttons.

Change-Id: I1c73f901344e21bd62e67890ce544f644c0cdda6
Reviewed-on: https://code.wireshark.org/review/17898
Reviewed-by: Gerald Combs <gerald@wireshark.org>
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>