Qt: valign text in RTP Stream Analysis window.
The text on the left hand side of the RTP Stream Analtsis window is
a (HTML) table of table headers followed by table data. For some reason
the table data is center aligned in the cells of the table. This does not
line up with the text in the table header cells.

See http://imgur.com/a/QIQFd

Also the HTML document is not compliant. Closing tags are missing, for
td and html itself.

Change-Id: I1cd58e8702db12433c863757141a6b68c1e2550c
Reviewed-on: https://code.wireshark.org/review/18074
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Gerald Combs <gerald@wireshark.org>