Add support for 6P-sixtop protocol in the IEEE802.15.4 dissector.
This patch adds one more Payload Information Element called 6P or sixtop
defined by the IETF 6TiSCH Working Group. Here it is shown the following
6P-sixtop information:
	-The 6P-sixtop version.
	-The 6P-sixtop code.
	-The 6P-sixtop Scheduling Function ID.

Then, depending of the 6P-sixtop code, a specific information is shown.
The supported codes -commands- for this patch are:
	-Add.
	-Delete.
	-Count.
	-List.
	-Clear.
	-Return Code Success.

For more information about the 6P-sixtop protocol, go to:

https://tools.ietf.org/html/draft-ietf-6tisch-6top-protocol-02

Bug 13005

Change-Id: I46b015dd8acd1515bff4090c5b39ad3897606f67
Reviewed-on: https://code.wireshark.org/review/18151
Reviewed-by: Michael Mann <mmann78@netscape.net>